+++ src/core-metadata/src/test/java/org/apache/kylin/measure/bitmap/BitmapCounterTest.java
@@ -22,7 +22,6 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
- 
 /*
  * Licensed to the Apache Software Foundation (ASF) under one
  * or more contributor license agreements.  See the NOTICE file
@@ -43,11 +42,11 @@
 
 package org.apache.kylin.measure.bitmap;
 
-import org.junit.Test;
-
 import static org.junit.Assert.assertArrayEquals;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
+
+import org.junit.Test;
 
 public class BitmapCounterTest {
     private static final BitmapCounterFactory factory = RoaringBitmapCounterFactory.INSTANCE;
@@ -68,14 +67,14 @@ public class BitmapCounterTest {
 
         counter2.orWith(counter);
         assertEquals(4, counter.getCount());
-        assertEquals(6, counter2.getCount());  // in-place change
+        assertEquals(6, counter2.getCount()); // in-place change
 
         int i = 0;
-        int[] values = new int[(int) counter2.getCount()];
-        for (int value : counter2) {
+        long[] values = new long[Math.toIntExact(counter2.getCount())];
+        for (long value : counter2) {
             values[i++] = value;
         }
-        assertArrayEquals(new int[]{10, 20, 30, 40, 1000, 2000}, values);
+        assertArrayEquals(new long[] { 10, 20, 30, 40, 1000, 2000 }, values);
 
         counter2.clear();
         assertEquals(0, counter2.getCount());
